embedding-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
